Move get_base_url
