+++ app/routes/session.py
@@ -13,7 +13,8 @@ from werkzeug.exceptions import Unauthorized
 from app.authentication.authenticator import decrypt_token, store_session
 from app.authentication.jti_claim_storage import JtiTokenUsed, use_jti_claim
 from app.globals import get_session_store, get_session_timeout_in_seconds
-from app.helpers.template_helpers import get_base_url, render_template
+from app.helpers import get_base_url
+from app.helpers.template_helpers import render_template
 from app.storage.metadata_parser import (
     validate_questionnaire_claims,
     validate_runner_claims,
